move runtimelib message bits into jupyter-protocol
