+++ crates/jupyter-protocol/Cargo.toml
@@ -10,6 +10,8 @@ license = "BSD-3-Clause"
 jupyter-serde = { path = "../jupyter-serde", version = "0.4.0" }
 async-trait = { workspace = true }
 anyhow = { workspace = true }
+bytes = { workspace = true }
+chrono = { workspace = true }
 futures = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
