+++ crates/jupyter-protocol/src/lib.rs
@@ -2,6 +2,8 @@ pub use jupyter_serde::*;
 pub mod messaging;
 pub use messaging::*;
 
+mod time;
+
 use async_trait::async_trait;
 use futures::{Sink, Stream};
